requirements/base.txt
django>=5.0.7
psycopg2-binary==2.9.9
djangorestframework==3.15.2
markdown==3.6
django-filter==24.3
requests==2.32.3
